Update symlink script

+++ symlink-dotfiles.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-rm -rdf $HOME/.gitconfig
+[ -f $HOME/.gitconfig ] && rm -rdf $HOME/.gitconfig
 ln -s $HOME/.dotfiles/.gitconfig $HOME
 ln -s $HOME/.dotfiles/.bash_aliases $HOME
 ln -s $HOME/.dotfiles/.tmux.conf $HOME
